Refactor tests: organize fixtures and utilities
- Move tests/fixtures/ to tests/integration/fixtures/
- Create tests/utils/ with helpers (files.py, plots.py, jobs.py)
- Add fixtures with docstrings: paths, config, data, jobs
- Update conftest.py imports and integration test paths
- All 54 tests pass
