+++ tests/integration/fixtures/config/templates/domain_gap.yaml
