+++ tests/integration/fixtures/config/templates/representativeness.yaml
