+++ tests/integration/fixtures/config/templates/visual_features.yaml
