+++ tests/integration/fixtures/data/challenge_welding.csv
